Implement full-text search be leveraging Elasticsearch (#19)
The search functionality of Ptolemy is now powered by the Elasticsearch
engine. All the features from the querying language are now implemented
and should be working properly. There's still room for optimization for
the Elasticsearch queries but it works well enough for now.

An Elasticsearch single-node cluster has been added to the development
docker-compose configuration, in order to ease contributions to the
project.

Connexion to the Elasticsearch cluster is configured through the
following environment variables:

* `ELASTICSEARCH_HOST` (defaults to `"localhost"`): the hostname or IP
of the node you want to connect to.
* `ELASTICSEARCH_PORT` (defaults to `9200`): the port on which the node
you want to connect to is listening.
* `ELASTICSEARCH_USER` (defaults to `nil`): the username to use when
connecting to the node.
* `ELASTICSEARCH_PASS` (defaults to `nil`): the password to use when
connecting to the node.