Authentication system (#21)
For now, it only exposes a traditional email/password way of signing in
but it supports having as many provider of authentication as needed. A
provider can be anything that needs an identifier and a token in order
to identify a user. For the email provider, the uid column stores the
email used to sign in, and the token column stores a Pbkdf2 hash of the
password chosen by the user.