User invites (#27)
This PR adds the ability for users to invite other users. To invite a user, you use an email so that Ptolemy can send them a link (expiring after 5 days) allowing them to register. The registration form allow them to set their email and their password in order to create an account. Once the account is created, the invite link gets invalidated. The user table also stores the user that invited them, but this information is not exposed as of yet.

Fixes #20