Use runtime configuration for Plug.Session (#41)
The System.get_env used to configure Plug.Session were being evaluated at
compile-time, rendering them completely useless in a release-context. This PR
introduces a new TequilaWeb.SessionPlug, wrapping the behavior of the original
Plug.Session, while allowing configuration compatible with the release
mechanism. The plug options must be overriden inside of the call/2 callback,
due to the fact that the init/1 callback is evaluated at compile-time,
exhibiting the same problem that we wanted to be fixed.